requirements.txt
lumibot
alpaca_trade_api
pandas
yfinance
numpy
matplotlib
python-dotenv
